projects/invoice-assistant/backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
anthropic==0.52.0
pdfplumber==0.11.6
python-dotenv==1.1.0